requirements.txt
sounddevice
numpy
faster-whisper
torch
torchaudio
anthropic
pynput
pystray
Pillow
